Move arrow key navigation to custom hook
