+++ src/components/molecules/Combobox/Listbox/Listbox.tsx
@@ -1,6 +1,6 @@
 import type { ListboxProps } from "./type";
 import { ListElement } from "./ListElement";
-import { useEffect, useRef } from "react";
+import { useListArrowNavigation } from "../hooks";
 
 export const Listbox = ({
   options,
@@ -10,33 +10,10 @@ export const Listbox = ({
   visualFocusIndex,
   onMouseOver,
 }: ListboxProps) => {
-  const listRef = useRef<HTMLUListElement>(null);
-  const listElementRefs = Array.from({ length: options.length }, () => useRef<HTMLLIElement>(null));
-
-  useEffect(() => {
-    if (!listRef.current || visualFocusIndex === null) {
-      return;
-    }
-
-    const currentFocusedElement = listElementRefs[visualFocusIndex].current;
-
-    if (!currentFocusedElement) {
-      return;
-    }
-
-    const listPosition = listRef.current.getBoundingClientRect();
-    const listElementPosition = currentFocusedElement.getBoundingClientRect();
-
-    if (
-      listElementPosition.bottom > listPosition.bottom ||
-      listElementPosition.top < listPosition.top
-    ) {
-      currentFocusedElement.scrollIntoView({
-        behavior: "smooth",
-        block: "nearest",
-      });
-    }
-  }, [listRef, visualFocusIndex, listElementRefs]);
+  const { listRef, listElementRefs } = useListArrowNavigation({
+    visualFocusIndex,
+    length: options.length,
+  });
 
   return (
     <ul
