+++ src/components/molecules/Combobox/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from "./useKeyboard";
 export * from "./useOutsideClick";
+export * from "./useListArrowNavigation";
